re-implement the STM32 flash
I started with sending "Flash Flash" command to the Pi from Qt. On reception, Pi would execute openocd command using `system()`. However, this worked when I run the Pi software by ssh but not when the software executes on startup. The fault could be my incapability to debug the problem running on two different platforms.

On discussion with Saurav, I was convinced that is ssh works then why not automate flash with ssh! I did exactly that and it works now!