add field for local port
I could not establish UDP connection using two ground stations running on same desktop. The issue turned out to be the fact that both of the GS had same local port (8081) trying to connect to respective server (corresponding Rpi). This issue is sorted by simply allowing a field to enter an unique port number to a satellite.